backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
passlib[bcrypt]
python-jose[cryptography]
